move tests in each packages
