+++ examples/psap/index.html
@@ -15,7 +15,6 @@
           <button class="pause">pause</button>
           <button class="stop">stop</button>
         </div>
-
         <div class="wrapper">
           <div class="ball"></div>
         </div>
