+++ packages/interpol/tests/Interpol.delay.test.ts
@@ -1,6 +1,6 @@
 import { it, expect, describe, vi } from "vitest"
-import { Interpol } from "../packages"
 import { wait } from "./utils/wait"
+import { Interpol } from "../src"
 
 describe.concurrent("Interpol delay", () => {
   it("play with delay", () => {
