+++ packages/interpol/tests/Interpol.refresh.test.ts
@@ -1,7 +1,8 @@
 import { it, expect, describe, vi } from "vitest"
-import { Interpol } from "../packages"
 import { randomRange } from "./utils/randomRange"
+import { Interpol } from "../src"
 import { wait } from "./utils/wait"
+
 
 describe.concurrent("Interpol refresh", () => {
   it("should compute 'from' 'to' and 'duration' if there are functions", async () => {
