+++ packages/interpol/tests/Timeline.play.test.ts
@@ -1,5 +1,5 @@
 import { it, expect, vi, describe } from "vitest"
-import { Timeline, Interpol, Ease } from "../packages"
+import { Timeline, Interpol, Ease } from "../src"
 
 describe.concurrent("Timeline play", () => {
   it("Timeline should add Interpol's and play properly", () => {
