+++ packages/psap/tests/psap.specificProperties.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from "vitest"
 import { JSDOM } from "jsdom"
-import { psap } from "../packages"
+import { psap } from "../src"
 import { randomRange } from "./utils/randomRange"
-import { validTransforms } from "../packages/psap/src/psap"
+import { validTransforms } from "../src/psap"
 
 const getDocument = () => {
   const dom = new JSDOM()
