+++ packages/psap/tests/utils/wait.ts
@@ -1,0 +1,1 @@
+export const wait = async (t) => new Promise((r) => setTimeout(r, t))
