+++ packages/psap/vite.config.ts
@@ -17,7 +17,6 @@ export default defineConfig({
       entry: [resolve(__dirname, "src/index.ts")],
       name: "psap",
       fileName: "psap",
-      formats: ["es", "cjs"],
     },
   },
